reorganised files in the toolkit folder
