+++ pyproject.toml
@@ -22,10 +22,10 @@ packages = ["src/goose"]
 goose-ai = "goose.module_name"
 
 [project.entry-points."goose.toolkit"]
-developer = "goose._internal.toolkit.developer:Developer"
-github = "goose._internal.toolkit.github:Github"
-screen = "goose._internal.toolkit.screen:Screen"
-repo_context = "goose._internal.toolkit.repo_context.repo_context:RepoContext"
+developer = "goose._internal.toolkit:Developer"
+github = "goose._internal.toolkit:Github"
+screen = "goose._internal.toolkit:Screen"
+repo_context = "goose._internal.toolkit:RepoContext"
 
 [project.entry-points."goose.profile"]
 default = "goose._internal.profile:default_profile"
