+++ src/goose/_internal/exchange/system.jinja
