+++ src/goose/_internal/exchange/view.py
