+++ src/goose/_internal/toolkit/__init__.py
@@ -1,3 +1,12 @@
+from .developer.developer import Developer  # noqa: F401
+from .repo_context.repo_context import RepoContext  # noqa: F401
+from .summarization.summarize_repo import SummarizeRepo  # noqa
+from .summarization.summarize_project import SummarizeProject  # noqa
+from .summarization.summarize_file import SummarizeFile  # noqa: F401
+from .github import Github  # noqa: F401
+from .screen import Screen  # noqa: F401
+
+
 from functools import cache
 
 from goose.toolkit import Toolkit
