+++ src/goose/_internal/toolkit/developer/prompts/developer.jinja
