+++ src/goose/_internal/toolkit/developer/prompts/safety_rails.jinja
