+++ src/goose/_internal/toolkit/developer/utils/check_shell_command.py
