+++ tests/_internal/toolkit/developer/test_check_shell_command.py
@@ -1,5 +1,5 @@
 import pytest
-from goose.utils.check_shell_command import is_dangerous_command
+from goose._internal.toolkit.developer.utils.check_shell_command import is_dangerous_command
 
 
 @pytest.mark.parametrize(
